+++ tests/com/splunk/RoleTest.java
@@ -108,10 +108,4 @@ public class RoleTest extends SDKTestCase {
             role.remove();
         }
     }
-    
-    // === Utility ===
-    
-    private void assertEquals(String[] a1, String[] a2) {
-        assertEquals(Arrays.asList(a1), Arrays.asList(a2));
-    }
 }