+++ tests/com/splunk/SDKTestCase.java
@@ -263,6 +263,10 @@ public abstract class SDKTestCase extends TestCase {
         return Arrays.asList(array).contains(value);
     }
     
+    protected static void assertEquals(String[] a1, String[] a2) {
+        assertEquals(Arrays.asList(a1), Arrays.asList(a2));
+    }
+
     protected static String locateSystemLog() {
         final String filename;
         String osName = service.getInfo().getOsName();